+++ League planner/IncidenciasPage.xaml.cs
@@ -30,7 +30,8 @@ namespace League_planner
         List<object> incidenciasVisitante = new List<object>();
 
         int golesLocal=0, golesVisitante=0;
-
+        int IDVisitante;
+        int IDLocal;
 
         public IncidenciasPage(Page previous, Gol gol)
         {
@@ -72,6 +73,10 @@ namespace League_planner
         {
             if(Eventos.SelectedItem != null)
             {
+                ILocal.ItemsSource = null;
+                IVisitante.ItemsSource = null;
+                incidenciasLocal.Clear();
+                incidenciasVisitante.Clear();
                 Calendario evento = (Eventos.SelectedItem as Calendario);
                 local.Content = evento.local;
                 visitante.Content = evento.visitante;
@@ -94,6 +99,9 @@ namespace League_planner
                         idVisitante = equipos[i].id;
                     }
                 }
+                IDVisitante = (int)idVisitante;
+                IDLocal = (int)idLocal;
+
                 //MessageBox.Show("id equipo -> " + idLocal);
                 clave = Convert.ToInt32(idLocal);
                 jugadorLocal.ItemsSource = App.JugadorController.GetAll().FindAll(x => x.Equipo==idLocal);
@@ -101,8 +109,27 @@ namespace League_planner
                 jugadorVisitante.ItemsSource = App.JugadorController.GetAll().FindAll(x => x.Equipo==idVisitante);
                 jugadorVisitante_Copy.ItemsSource= App.JugadorController.GetAll().FindAll(x => x.Equipo == idVisitante);
 
+                
+                List <Gol> goles=App.GolController.GetAll().FindAll(z => z.clave_partido == evento.id);
+                List<Tarjeta> tarjetas = App.TarjetaController.GetAll().FindAll(t => t.Partido == evento.id);
+
+
+
+                incidenciasLocal.AddRange( goles.FindAll(x => x.equipo == IDLocal));
+                golesLocal = incidenciasLocal.Count;
+                golesL.Content = golesLocal;
+                incidenciasLocal.AddRange(tarjetas.FindAll(t=>t.equipo==IDLocal));
+
+                incidenciasVisitante.AddRange(goles.FindAll(x => x.equipo == IDVisitante));
+                golesVisitante = incidenciasVisitante.Count;
+                golesV.Content = golesVisitante;
+                incidenciasVisitante.AddRange(tarjetas.FindAll(x => x.equipo == IDVisitante));
+
+                
                 ILocal.ItemsSource = incidenciasLocal;
                 IVisitante.ItemsSource = incidenciasVisitante;
+
+
 
             }
         }
@@ -179,13 +206,18 @@ namespace League_planner
             {
                 if (j != null)
                 {
-                    Gol g = new Gol((int)j.id, (int)c.id, 0);
+                    Gol g = new Gol((int)j.id, (int)c.id, 0, IDLocal);
                     incidenciasLocal.Add(g);
                     ILocal.ItemsSource = null;
                     ILocal.ItemsSource = incidenciasLocal;
                     App.GolController.Save(g);
                     golesLocal++;
-                    golesL.Content = golesLocal;                }
+                    golesL.Content = golesLocal;
+                    updateWinner();
+
+
+                }
+
                 else
                     MessageBox.Show("Seleccione un jugador para agregar un gol");
             }
@@ -207,7 +239,7 @@ namespace League_planner
                 {
                     if (!String.IsNullOrEmpty(tipoTLocal.Text))
                     {
-                        Tarjeta T = new Tarjeta((int)j.id, tipoTLocal.Text, (int)c.id);
+                        Tarjeta T = new Tarjeta((int)j.id, tipoTLocal.Text, (int)c.id,IDLocal);
                         incidenciasLocal.Add(T);
                         ILocal.ItemsSource = null;
                         ILocal.ItemsSource = incidenciasLocal;
@@ -233,13 +265,14 @@ namespace League_planner
             {
                 if (j != null)
                 {
-                    Gol G = new Gol((int)j.id, (int)c.id, 0);
+                    Gol G = new Gol((int)j.id, (int)c.id, 0,IDVisitante);
                     incidenciasVisitante.Add(G);
                     IVisitante.ItemsSource = null;
                     IVisitante.ItemsSource = incidenciasVisitante;
                     App.GolController.Save(G);
                     golesVisitante++;
                     golesV.Content = golesVisitante;
+                    updateWinner();
                 }
                 else
                     MessageBox.Show("Seleccione un jugador para agregar un gol");
@@ -260,7 +293,7 @@ namespace League_planner
                 {
                     if (!String.IsNullOrEmpty(tipoTVisitante.Text))
                     {
-                        Tarjeta T = new Tarjeta((int)j.id, tipoTVisitante.Text, (int)c.id);
+                        Tarjeta T = new Tarjeta((int)j.id, tipoTVisitante.Text, (int)c.id,IDVisitante);
                         incidenciasVisitante.Add(T);
                         IVisitante.ItemsSource = null;
                         IVisitante.ItemsSource = incidenciasVisitante;
@@ -281,25 +314,30 @@ namespace League_planner
         {
             if (IVisitante.SelectedItem != null)
             {
-                if (IVisitante.SelectedItem.GetType() == typeof(Gol))
+
+                if (MessageBox.Show("Esta seguro que quiere eliminar Este Elemento?", "Confirmar eliminación", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+
                 {
-                    Gol g = (IVisitante.SelectedItem as Gol);
-                    App.GolController.Delete(g);
-                    golesVisitante--;
-                    golesV.Content = golesVisitante;
+                    if (IVisitante.SelectedItem.GetType() == typeof(Gol))
+                    {
+                        Gol g = (IVisitante.SelectedItem as Gol);
+                        App.GolController.Delete(g);
+                        golesVisitante--;
+                        golesV.Content = golesVisitante;
+                        updateWinner();
+                    }
+
+                    if (IVisitante.SelectedItem.GetType() == typeof(Tarjeta))
+                    {
+                        Tarjeta t = (IVisitante.SelectedItem as Tarjeta);
+                        App.TarjetaController.Delete(t);
+                    }
+
+                    incidenciasVisitante.Remove(IVisitante.SelectedItem);
+                    IVisitante.ItemsSource = null;
+                    IVisitante.ItemsSource = incidenciasVisitante;
+
                 }
-
-                if (IVisitante.SelectedItem.GetType() == typeof(Tarjeta))
-                {
-                    Tarjeta t = (IVisitante.SelectedItem as Tarjeta);
-                    App.TarjetaController.Delete(t);
-                }
-
-                incidenciasVisitante.Remove(IVisitante.SelectedItem);
-                IVisitante.ItemsSource = null;
-                IVisitante.ItemsSource = incidenciasVisitante;
-               
-
             }
         }
 
@@ -307,25 +345,46 @@ namespace League_planner
         {
             if (ILocal.SelectedItem != null)
             {
-                if (ILocal.SelectedItem.GetType() == typeof(Gol))
-                {
-                    Gol g = (ILocal.SelectedItem as Gol);
-                    App.GolController.Delete(g);
-                    golesLocal--;
-                    golesL.Content = golesLocal;
-                }
+                if (MessageBox.Show("Esta seguro que quiere eliminar Este Elemento?", "Confirmar eliminación", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
 
-                if (ILocal.SelectedItem.GetType() == typeof(Tarjeta))
                 {
-                    Tarjeta t = (ILocal.SelectedItem as Tarjeta);
-                    App.TarjetaController.Delete(t);
-                }
+                    if (ILocal.SelectedItem.GetType() == typeof(Gol))
+                    {
+                        Gol g = (ILocal.SelectedItem as Gol);
+                        App.GolController.Delete(g);
+                        golesLocal--;
+                        golesL.Content = golesLocal;
+                        updateWinner();
+                    }
 
-                incidenciasLocal.Remove(ILocal.SelectedItem);
-                ILocal.ItemsSource = null;
-                ILocal.ItemsSource = incidenciasLocal;
+                    if (ILocal.SelectedItem.GetType() == typeof(Tarjeta))
+                    {
+                        Tarjeta t = (ILocal.SelectedItem as Tarjeta);
+                        App.TarjetaController.Delete(t);
+                    }
+
+                    incidenciasLocal.Remove(ILocal.SelectedItem);
+                    ILocal.ItemsSource = null;
+                    ILocal.ItemsSource = incidenciasLocal;
+                }
                 
             }
+        }
+
+        private void updateWinner()
+        {
+
+            Calendario evento = (Eventos.SelectedItem as Calendario);
+            if (golesLocal > golesVisitante)   
+                evento.ganador = IDLocal;
+            else if (golesLocal < golesVisitante)
+                evento.ganador = IDVisitante;
+            else
+                evento.ganador = -1;
+
+
+            App.CalendarioController.Save(evento);
+
         }
     }
 }
